[FIX] hr_employee_calendar_planning: Change test employee name
This test collides with the dms module. In this test an employee is
being created with a name that already exists in another test.
The employee's file directory is created with the name of this employee.
As there is already an employee with that name, there is already a
directory with that name and this is what causes the error. As the
probability of 2 employees with the same name coinciding in the same
company is very small, instead of intervening in the logic for the
creation of the employee's directory, we change the name of the
employee in the test so that it does not cause the error when
executing this module together with dms.
In the hypothetical case of two employees with the same name within
the same company, we would have to resort to using other tools, either
by creating an additional module to modify the logic and add an employee
identifier to the name of the directory.

TT52110